Reuse `client.auth` to retrieve token
