+++ vng_api_common/client.py
@@ -87,26 +87,3 @@ def get_client(url: str) -> Client | Any | None:
         return None
 
     return build_client(service.api_root, client_factory=client_class)
-
-
-def get_auth_headers(
-    client_id: str,
-    client_secret: str,
-    user_id: str = "",
-    user_representation: str = "",
-    **claims,
-) -> dict:
-    payload = {
-        # standard claims
-        "iss": client_id,
-        "iat": int(time.time()),
-        # custom claims
-        "client_id": client_id,
-        "user_id": user_id,
-        "user_representation": user_representation,
-        **claims,
-    }
-
-    encoded = jwt.encode(payload, client_secret, algorithm=JWT_ALG)
-
-    return {"Authorization": "Bearer {encoded}".format(encoded=encoded)}
